0# This is a combination of 8 commits.
new chapter list enbeded in markdown readme,
chapter files renumbered to respct ordering.
chapter 11 error when trying to render in liascript.
A description for each chapter was added to a better overview.

new links for chapter. chapter 11 is not rendered in liascript. error in the rendering

adding description of each chapter

moving chapter link emoji out of the title. adding link to chapter intitle and bellow text

adding space between image and text